add Readme and extract the process of loading RSchemaRegion to RSchemaRegionLoader
